requirements.txt
opencv-contrib-python~=4.9.0.80
numpy~=1.26.4
pandas~=2.2.2
Pillow~=10.4.0
dlib-bin; platform_system != "Windows"  # optional (Linux/macOS)
matplotlib~=3.9.0
fastapi~=0.115.0
uvicorn[standard]~=0.30.0
python-dotenv~=1.0.1
SQLAlchemy~=2.0.32
alembic~=1.13.2
bcrypt~=4.1.3
PyJWT~=2.9.0
reportlab~=4.2.2
openpyxl~=3.1.5
pytest~=8.3.2
